Changed backdrop-filter polyfill strategy from child to sibling
Although adding the blurred overlay as a child results in easier coding for the positioning and shaping, it renders over the background and requires weird z indices. For that reason, a node is added directly before the polyfilled node, where it will be rendered under the original node, and its positioning is replicated. This way, the background rendering will be unaffected.