Added autogenerated testing skeleton
Since no tests are currently included, and the project is already quite
large in size, it is unfeasible to write all unit tests from scratch, by
hand. Therefore, I've autogenerated a testing skeleton using phpunitgen
and modified it to integrate with Symfony testing structures. These
tests are stubs and still need to be fixed or implemented, but it
provides a simple structure to begin with for upcoming commits.