Fix invalid creation and testing of values/objects in unit tests
In a number of unit tests, a general value assignment was generated.
These were invalid for numerous reasons (protected constructor, objects
accessed as a service, etc.). In a number of cases, default values would
be added by the tested class as well, resulting in failing tests. Both
cases are fixed by this commit.