feat(ical): kiwi will send an ical email and provide public ical endpoint
There are currenty different methods for a single event and the public
feed. In the future we should see if we can merge these.

Currently during testing the public url could only be added using http,
most likely this is a resullt from the tunnel I used to expose my local
build.