Restrict activities in the open iCal URL
With the new iCal URL, all activities were embedded in the feed.
However, this included activities that are not visible when not logged
in. This affects activities that have a target audience assigned.

This commit intends to fix that behaviour, by only retrieving the 
activities from the database that have no target audience assigned to
them.