Add TestData builder class
This commit adds the TestData helper class. The class is intended to
generate many possible permutations of data. This is done by emulating a
factory-like structure, where multiple possible values can be given for
a specified data field. Moreover, test conditions may be given that can
evaluate and reject impossible combinations of values. Upon generation,
all possible permutations of data will be generated. This results in
easier and broader testing.