Initial example of data fixture
This commit adds the DoctrineFixtureBundle, the DAMADoctrineTestBundle
and the LiipTestFixturesBundle to the project. The first bundle is to
add default (testing) data. The second bundle is to speed up testing, by
not actually committing changes to the database. The last bundle is to
allow access to this data from tests. Moreover, an initial example of a
fixture has been added for activity data.