Added feature for hidden activities
When creating or editing an activity, the organiser will now be given the choice to set the activity to Hidden. For this I added a checkbox in the ActivityEditType and ActivityNewType files. The organiser will also be given the option whether or not users should be mailed. So I also added an additional check in the RegistrationController. Lastly it also requires an additional column in the database to keep track. So another activity variable was added in the Activity entity