Initial refactoring of update.php
This commit intends to simplify and refactor the update.php script. Initial steps were taken to deduplicate code. Moreover, shorthands were
introduced where possible. Moreover, the CreateLocalAccountCommand was
modified to be used instead of the CreateLocalAdminCommand, which was a
copy. Although untested, these changes should not modify the behaviour
of the code.