Remove redundant versions of installer
This commit removes the install.ini file and the kiwiinstaller.php file.
The install.ini file is replaced by the .env.local.php, which also
interfaces with Kiwi itself. The kiwiinstaller.php file is replaced by
the update.php file, which provides a GUI from the browser and is
therefore more user-friendly.