Remove assets when serving in testing
During testing, the building of JS files is redundant and costs time.
Therefore, these steps are not present in the testing.yml workflow.
However, when serving, Twig expected these files to be present. This
commit changes that behaviour. Twig won't look for built JS now and will
simply continue on.