Move fixtures to tests/ directory
To support running functional tests, testing data is necessary. This was
accomplished using data fixtures, which were places in src/DataFixtures.
However, this results in them requiring code coverage. Since this code
is only used during testing, this is somewhat redundant. Therefore, they
have been moved to tests/Helper/Database, which is excluded from code
coverage computations. Note that the tests/Helper directory has been
added as a service; this is required for loading the fixtures, and may be convenient for the additional helper code as well.