Autoload test helper classes in production
Since services.yml specifies tests/Helper classes as services (necessary
for the fixtures), these classes are automatically loaded by Symfony.
However, they were not made available through the autoloader, which
results in a failing production workflow. Alternatively, a solution can
be implemented later to only load these classes as services in
development, but that solution seems to be more complex. Another
solution would be to move the fixtures to the src/ folder and annotate
them to be ignored for code coverage requirements, but this location
doesn't match their desired use. Therefore, this solution was chosen.