Branch Publish feature: adding comments to registrations
in this branch I added the feature for organisers/admins to add comments to registrations in the system. This required an additional datebase column.

The ability to add comments has been added to the present menu. and are displayed on the participant page of the activity. For this I added code to:

* present.html.twig, most changes are added here, the form fields are now in organised in tables.
* show.html.twig has been changed a minor amount to display the information
* presentType.php (form) been given an extra comment field.
* Registration has been given an additional parameter