Deletion of App\Provider\* and refactor of related code
This commit removes the src/Provider folder and it's contents. All code
that refers to it is refactored. Since it was used to aggregate Bunny
and local account data, this commit therefore currently breaks Bunny
integration. This should be fixed in a later commit. For that reason,
OAuth.. code has not been modified yet.