Commit for adding test coverage
Activity: Fixed bug where it had to return ?bool instead of bool.

ActivityControllerTest (activity): This was the one I had to write a test for. For this to work however, I had to implement the fixtures, but It for somereason refused to have fixtures in the Webtestcase, so i turned it into an Authwebtestcase which worked. Maybe someone needs to eventually look at this? For the rest its a normal test which looks at all non hidden activities, and if they show up.

ActivityControllerTest(admin): I was accidantly testing the wrong code, removed the code -> moved it to (activity folder).

ActivityEditTypeTest: Generic test case, creates a form and fills it with test data: see if it passes it.

ActivityNewTypeTest: same as ActivityEditTypeTest.

ActivityTest: Had to write test case for getHidden/ setHidden

ActivityRepositoryTest: had to write a repository test, so I had to load the fixtures. However loading fixtures was a fair amount of code. Maybe turn create a RepositoryTestCase that extends kernelTestCase which implements all this code instead of doing it per file? Test works rather simple: Get all the groups. then Execute the query search and see if the results are correct. (aka return atleast one value)