fix(dependencies): yarn packages are up to date
A lot of our yarn packages were already a little behind.
Now it became to the point where yarn couldn't compile anymore.

Since I was already updating dependencies I decided to update them all
to the latest stable version. Only thing I'm not sure if is if there is
a better place to make sass:math available, I now do it in each file
that needs is but perhaps making it globally available would be better.