feat(passwordcontroller): testing added
It might look like it's working, but it's hit or miss.
The tests fail ever so often. As far as I've seen Request never fails.

I assume the problem lay's in the token. and how the method get's it
from the request. further debugging is needed.

Also, this is happy flow testing. it should be cleaned up and extended.