Fix nasty ass bugs
This commit fixes a bug where the URL decoding on part of PHP itself
would sometimes misinterpret the password reset token. By encoding it in
a URL-safe manner, this behaviour is ommitted. Furthermore, a bug where
the OIDC parameters for the testLoginOidc() method would not be unset
and be dependent upon an external server are now mitigated.