Cleanup of GraphQL API
This commit cleans up the current state of the GraphQL API. It features
a renamed API route, which should provide more future flexibility.
Furthermore, the RootQuery has been renamed to Query, as is usually the
case in GraphQL APIs. Also, the Aggregate builder has been removed, as
it's currently not mature yet. Lastly, an initial functional test has
been added for the root queries.