Further expand GraphQL API access
Previously, the API calls were split into Query and AdminQuery files.
This reduced the exposal of data and is similar to the current access
policy in the interface. However, not all data is privacy-sensitive, this
commit intends to publicate more data. This is done by merging AdminQuery
with Query. The corresponding functional tests are updated accordingly.