Person data for current activities
As activities that are currently visible can show the names of registered
persons, this commit mimics that behaviour. A check has been added to
Activity which checks whether the activity is currently visible. This is
used to let GraphQL known whether the person field in Registration is
accessible. Additionally, fields in LocalAccount were further restricted
in their data access.