Authentication requirements of GraphQL API
This commit changes the authentication requirements for the activities
and groups queries for the API. Now, API users are required to be
authenticated, although not necessarily as admin. This is to accomodate
for users of Kiwi that are not necessarily interested in sharing all
their details with the world. The tests were expanded and updated for
this change. Note that, to allow for login after running loadFixtures
(for example outside of the setup method of a test), the AuthWebTestCase
login method had its loadFixtures() call removed. Other tests that relied
upon this behaviour were updated accordingly.