Add isAdmin() method to LocalAccount
This commit adds the isAdmin() method to the LocalAccount entity. An API
property and the corresponding test for it are added as well. Moreover,
all other API properties in the LocalAccount are protected by login now.
Although not strictly necessary given the current setup, I think it's
just good practice.