Hotfix for data override in Activity forms
Previously, a placeholder value was implemented in the activity forms for the
capacity field. However, this was done by overriding the data attribute in the
form, which modifies the actual data when an activity is editted. This is
unintended and can be easily overlooked by a user. This hotfix changes the
data attribute to the "attr" attribute, which directly integrates it into the
placeholder html attribute.