Groups can now only create activities for themselfs, or their child groups.
ActivityController: In the NewAction and EditAction, added code that gets an array of the groups it may use.

ActivityEditType: Added a group array variable in the options

ActivityNewType: Same as ActivityEditType

GroupRepository: Added 2 new repository functions. Get all childgroups of a group, and get all groups including childgroups of a person.

Known error: When a person from a group creates an activity, If this user selects the activity is not organised by a group, it returns a no access exception. However the activity is published