Add PhpStan static analyser to Kiwi
This commit adds the PhpStan static analyser to kiwi. It integrates a
number of additional extensions for Symfony, Doctrine and PHPUnit, and
provides a configuration file for PhpStan which tests both source code
and tests on the maximum analysis level. It has parallel processing
enabled for performance and integrates with the phpunit bridge by
Symfony, which provides phpunit outside of the vendor folder. The static
analyser has been integrated into the testing pipeline (as here, it's
guaranteed that phpunit was installed). Moreover, the linter has been
added for good measure as well.