Added phpstan extension installer as allowed composer plugin
Given the fact that I was using an older version of composer, I had no
issues regarding the extension installer. However, as the CI pipeline
was using the most up-to-date version, the fact that the explicit
allowing of the phpstan extension installer plugin wasn't added to
composer.json created a problem. I've now fixed this.