Database testing pipeline
This commit adds a github actions pipeline. The pipeline is intended to
check whether the migrations are matching the entity definitions. It
accomplishes this by running all migrations, after which it tries to
create a new migration automatically. If this fails, that means that no
changes are available, as intended. Then, it tries to load the fixtures
used in testing. After that, it tries to unmigrate to an empty database
as well, in order to check that the down migrations function properly as
well.