Improve integration between phpstan and kiwi
This commit adds additional integration between Kiwi and Phpstan. This
adds a number of new (higher quality) detected issues, and removes some
lower tier issues as well. Two files in the config are added that allow
for the integration, and the baseline file has been updated with the new issues as well.