Update GroupVoter to respect group inheritance
This commit updates the GroupVoter to respect inheritance. Previously, when a
user was in a parent group, they weren't "in" the child groups. However, this
is the behaviour that's expected, as child groups are managed by their parents,
which means that Kiwi should behave like they are in those groups. The code has
been updated to reflect this. Moreover, a group cannot manage itself.
Therefore, the 'edit_group' vote has been added, that only allows editing child
groups. Lastly, the 'any_group' vote has been dropped, as this was replaced by
the addition of the ROLE_AUTHOR role.