Send e-mails to participants
This commit adds a number of buttons for admins & organisers, which allows them
to send e-mails to participants for activities. A button is added for anyone
individually who has registered, deregistered or is on the reserve list.
Additionally, a button is added to send a mail to all registered participants,
which respects the registrants privacy by using the BCC field. The system works
by utilizing the "mailto:" URL construct.