E-mails to participants in new window
This commit adds a target="_blank" to all added 'mailto:' links in admin &
organise. When the mail handler is set to a browser-based mail application,
this makes sure a new tab is opened. Note that this doesn't work in firefox due
to a long-standing [bug](https://bugzilla.mozilla.org/show_bug.cgi?id=646552).