Merge RegistrationHelper and add registration events
This commit is a refactor of the registration related controller code.
Previously, that part of the codebase had a RegistrationHelper for the code
that was both related to the admin and organise functionality. With the removal
of the organise module, the code was simplified by merging that code into the
RegistrationController in admin. Moreover, events were added for new and
removed registrations, which makes the code both more simple and extendable.