Add deduplication form for locations
Since locations are automatically created, a lot of duplicate locations might
have been added for environments that have been in production for some time.
Moreover, users might be more careless when working with locations, as they are
not the primary "unit" of Kiwi. This commit adds options to deduplicate
locations.