Remove standard group arrangement
This commit removes the standard group arrangement. This was done, because this
branch contains a lot of functionality that cannot be utilized with the old
setup. Therefore, a migration has been added to archive the old data groups, if
any groups are present. This allows admins to easily migrate to a fresh setup,
while for new setups, it doesn't make a difference.