Initial impl. for external registrants
This commit serves as the initial implementation for participants at an
activity that do not have a registered user account within the system. An
(embedded) class has been added to the Registration entity, that provides user
information if no LocalAccount instance is referred, called ExternalRegistrant.
This class shares a new ContactInterface with the LocalAccount class. A button
has been added to the admin, and the RegistrationType form has been updated
accordingly. When relevant code assumed a LocalAccount instance, modifications
have been made to support the new ExternalRegistrant instance as well.