External registrants are removed instead of deletedate
With normal registrations for activities, the removal of a registration constitutes the
value 'deleteDate' of a registration to be set from null to the current date/time. For
external registrants however, their data cannot be retained. Therefore, their data must
be removed immediately. This allows to resolve a bug when visualizing deregistrations
in the admin as well.