Hide personal details from external registrants in the frontend
This commit changes the way registrants are shown in the frontend. Previously, both
internal and external registrants were shown in the frontend. However, since external
registrants don't have access to kiwi, they don't have access to the list of
registrants, while users with an account do have access to their personal details. This
commit changes that: only personal details of persons with an account are shown. This
removes the asymmetry of information access.