Hi,
After reviewing the comments on my first PR I made some changes.

"The form is responsible for updating any value, but if (in the future) details or the confirmation message are added to the form, this code will break. Therefore, please move these 3 lines to ln. 109, as this will improve the resiliency of the code" This has been done and everything still seems to be working!

"This code checks whether an e-mail already exist, and therefore "validates" the user input. To do this, we use validation constraints instead. Please review the UniqueEntity constraint to see how to implement this"
A validation constrained has been added. Whenever a user tries to make an account with an already existing email they'll get the message: "This value is already used."

"AddOption is called here, but I'm assuming you mean addOption? Please correct this" This has been corrected.

And all the latest changes should be merged into this PR!