Always show an author selection
Previously, the author field was hidden when the user was only member of a single
active group. However, that group was not persisted with the newly created activity,
which causes the person that created the activity to be unable to view or edit it; only
admins can see it. This commit simplifies the flow by always showing the selection box.
The browser automatically sets a default option.