Hotfixes 2023-01-09
This commit contains a number of hotfixes. Firstly, creating and editing of activities
was fixed by editing a query in th ActivityEditType. The error only occurs in
deployment, which suggests that it might be related to MariaDB usage instead of MySQL.
Furthermore, presence registration was fixed when any external registrants are present
by adding a __toString() implementation. Lastly, the overview of groups for an account
was fixed, as this still refered to the removed getRoot method.