Hi,
In this PR I made an extra feature for making activities.

It is now possible for activities to be saved as a template ("sjabloon"). This can then be used to make new activities. This feature can be usefull, for when there are repeating activities.

Could someone take a look at this?